Move map popup into separate file
